% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchingCorrelatedRandomWalk.R
\encoding{UTF-8}
\name{scrw}
\alias{scrw}
\title{Function for Estimating Correlated Random-Walk from Tracking Data}
\usage{
scrw(
  data,
  mode = FALSE,
  obs_error_type = "normal",
  overdispersion = FALSE,
  minDF = 2,
  gamma_fix = rep(-1, as.numeric(mode) + 1),
  theta_fix = rep(-1, as.numeric(mode) + 1),
  sigma_key = c(0, 1),
  mixture = TRUE,
  logit_p_init = -0.5,
  phase = 1
)
}
\arguments{
\item{data}{tracking data with the first column of longitude (easting) and latitude (northing)}

\item{mode}{whether different modes (migration and staying) is assumed (TRUE) or not (FALSE; default)}

\item{obs_error_type}{observation error distribution "normal" (default) or "t" (NOT work at present)}

\item{overdispersion}{whether overdispersion in binomial distribution is estimated (TRUE; may not walk) or not (FALSE; default)}

\item{minDF}{minimum df parameter in t-distribution}

\item{gamma_fix}{set fixed gamma value for each mode if neccessary (e.g., c(1,0))}

\item{theta_fix}{set fixed theta value for each mode if neccessary (e.g., c(0,0))}

\item{sigma_key}{able to assume the same observation error bewteen longitude and latitude by using c(0,0)}

\item{mixture}{whether using mixture of two modes}

\item{logit_p_init}{initial value for logit_p}

\item{phase}{the number of phase estimation when mode=TRUE}
}
\value{
list of elements below
\describe{
\item{\code{input}}{used argument}
\item{\code{opt}}{\code{nlminb} object}
\item{\code{MakeADFun}}{\code{TMB::MakeADFun} object}
\item{\code{rep}}{\code{TMB::sdreport} object}
\item{\code{par_list}}{list of all parameters}
\item{\code{x}}{estimated location}
\item{\code{p}}{probability of migration mode}
\item{\code{alpha}}{probability of the same mode (migration or staying) at a next step}
\item{\code{phi}}{process error of binomial distribution (fixed at \code{phi=1} when \code{overdispersion=FALSE})}
\item{\code{det_p}}{deterministic probability of migration mode}
\item{\code{gamma}}{auto-corrlation coefficient for each mode; \code{0<=gamma<=1}}
\item{\code{theta}}{angle for each mode}
\item{\code{sigma}}{process error of random walk}
\item{\code{rho}}{correlation coefficient in process error between longitudinal and latitudinal directions}
\item{\code{omega}}{magnitude of observation error in normal distribution}
\item{\code{loglik}}{log-likelihood}
\item{\code{k}}{number of fixed parameters}
\item{\code{AIC}}{AIC}
\item{\code{AICc}}{AICc (sample size is assumed to be time-series x 2)}
\item{\code{BIC}}{BIC (sample size is assumed to be time-series x 2)}
}
}
\description{
Function for Estimating Correlated Random-Walk from Tracking Data
}
\examples{
\dontrun{
model = scrw(data,mode =TRUE,gamma_fix = c(1,0),theta_fix = c(-1,0))
}
}
