% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchingCorrelatedRandomWalk.R
\encoding{UTF-8}
\name{use_scrw_tmb}
\alias{use_scrw_tmb}
\title{Function for activating cpp file for TMB}
\usage{
use_scrw_tmb(
  TmbFile = "scrw",
  CppDir = system.file("executable", package = "SCRWtmb"),
  RunDir = getwd(),
  overwrite = FALSE
)
}
\arguments{
\item{TmbFile}{Cpp file name}

\item{CppDir}{directory having the cpp file}

\item{RunDir}{directory for running the code (default: working directory)}

\item{overwrite}{whether the cpp file is overwritten (TRUE) or not (FALSE; default)}
}
\description{
Function for activating cpp file for TMB
}
\examples{
\dontrun{
use_scrw_tmb()
}

}
