% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchingCorrelatedRandomWalk.R
\encoding{UTF-8}
\name{scrw}
\alias{scrw}
\title{Function for Estimating Correlated Random-Walk from Tracking Data}
\usage{
scrw(
  data,
  mode = FALSE,
  obs_error_type = "normal",
  overdispersion = FALSE,
  minDF = 2,
  gamma_fix = rep(-1, as.numeric(mode) + 1),
  theta_fix = rep(-1, as.numeric(mode) + 1),
  sigma_key = c(0, 1),
  mixture = TRUE,
  logit_p_init = -0.5,
  phase = 1
)
}
\arguments{
\item{data}{tracking data with the first column of longitude (easting) and latitude (northing)}

\item{mode}{whether different modes (migration and staying) is assumed (TRUE) or not (FALSE; default)}

\item{obs_error_type}{observation error distribution "normal" (default) or "t" (NOT work at present)}

\item{overdispersion}{whether overdispersion in binomial distribution is estimated (TRUE; may not walk) or not (FALSE; default)}

\item{minDF}{minimum df parameter in t-distribution}

\item{gamma_fix}{set fixed gamma value for each mode if neccessary (e.g., c(1,0))}

\item{theta_fix}{set fixed theta value for each mode if neccessary (e.g., c(0,0))}

\item{sigma_key}{able to assume the same observation error bewteen longitude and latitude by using c(0,0)}

\item{mixture}{whether using mixture of two modes}

\item{logit_p_init}{initial value for logit_p}

\item{phase}{the number of phase estimation when mode=TRUE}
}
\description{
Function for Estimating Correlated Random-Walk from Tracking Data
}
